After changing the loss to log(1-DG)
After changing the loss to log(1-DG), the discriminator was weak from the previous training so the gradients were high, but within 10 epochs it trained really fast, faster than the generator so the loss for Discriminator prediction became almost 0 so low gradients. there is a trade off to find here.
Now trying to make an average of both losses to see where that goes.